attempted some cleanup and restructuring
