+++ src/genh5_conversion/bindings.h
@@ -2,8 +2,22 @@
 #define GENH5_CONVERSION_BINDINGS_H
 
 #include "genh5_conversion/defaults.h"
-#include "genh5_datatype.h"
+
 #include "genh5_bytearray.h"
+#include "genh5_version.h"
+#include "genh5_reference.h"
+
+/** GENH5 **/
+namespace GenH5
+{
+
+GENH5_DECLARE_IMPLICIT_CONVERSION(GenH5::Version);
+GENH5_DECLARE_IMPLICIT_CONVERSION(GenH5::Reference);
+
+} // namespace GenH5
+
+GENH5_DECLARE_DATATYPE(GenH5::Version, DataType::Version);
+GENH5_DECLARE_DATATYPE(GenH5::Reference, DataType::Reference);
 
 /** STL **/
 GENH5_DECLARE_CONVERSION_TYPE(std::string, char*);
@@ -42,7 +56,17 @@ GENH5_DECLARE_DATATYPE(QString, DataType::VarString);
 #include <QVector3D>
 #include <QVector4D>
 
+namespace GenH5
+{
+
 GENH5_DECLARE_IMPLICIT_CONVERSION(QPoint);
+GENH5_DECLARE_IMPLICIT_CONVERSION(QPointF);
+GENH5_DECLARE_IMPLICIT_CONVERSION(QVector2D);
+GENH5_DECLARE_IMPLICIT_CONVERSION(QVector3D);
+GENH5_DECLARE_IMPLICIT_CONVERSION(QVector4D);
+
+} // namespace GenH5
+
 GENH5_DECLARE_DATATYPE_IMPL(QPoint)
 {
     using T = decltype (std::declval<QPoint>().x());
@@ -51,7 +75,6 @@ GENH5_DECLARE_DATATYPE_IMPL(QPoint)
     return GenH5::dataType<T, T>({"xp", "yp"});
 };
 
-GENH5_DECLARE_IMPLICIT_CONVERSION(QPointF);
 GENH5_DECLARE_DATATYPE_IMPL(QPointF)
 {
     using T = decltype (std::declval<QPointF>().x());
@@ -60,7 +83,6 @@ GENH5_DECLARE_DATATYPE_IMPL(QPointF)
     return GenH5::dataType<T, T>({"xp", "yp"});
 };
 
-GENH5_DECLARE_IMPLICIT_CONVERSION(QVector2D);
 GENH5_DECLARE_DATATYPE_IMPL(QVector2D)
 {
     using T = decltype (std::declval<QVector2D>().x());
@@ -69,7 +91,6 @@ GENH5_DECLARE_DATATYPE_IMPL(QVector2D)
     return GenH5::dataType<T, T>({"x", "y"});
 };
 
-GENH5_DECLARE_IMPLICIT_CONVERSION(QVector3D);
 GENH5_DECLARE_DATATYPE_IMPL(QVector3D)
 {
     using T = decltype (std::declval<QVector3D>().x());
@@ -78,7 +99,6 @@ GENH5_DECLARE_DATATYPE_IMPL(QVector3D)
     return GenH5::dataType<T, T, T>({"x", "y", "z"});
 };
 
-GENH5_DECLARE_IMPLICIT_CONVERSION(QVector4D);
 GENH5_DECLARE_DATATYPE_IMPL(QVector4D)
 {
     using T = decltype (std::declval<QVector4D>().x());
