+++ src/genh5_data/base.h
@@ -9,6 +9,7 @@
 #ifndef GENH5_DATA_BASE_H
 #define GENH5_DATA_BASE_H
 
+#include "genh5_conversion/defaults.h"
 #include "genh5_datatype.h"
 #include "genh5_dataspace.h"
