+++ src/genh5_datatype.h
@@ -11,10 +11,11 @@
 
 #include "genh5_idcomponent.h"
 #include "genh5_object.h"
-#include "genh5_conversion/type.h"
-#include "genh5_typetraits.h"
-#include "genh5_version.h"
-#include "genh5_utils.h"
+#include "genh5_typedefs.h"
+
+#include "H5Tpublic.h"
+
+#include <algorithm>
 
 namespace GenH5
 {
@@ -48,6 +49,8 @@ public:
     static DataType const& VarString;
 
     static DataType const& Version;
+
+    static DataType const& Reference;
 
     /// will create an string type with utf8 enabled by default
     static DataType string(size_t size, bool useUtf8 = true) noexcept(false);
@@ -186,137 +189,6 @@ getTypeNames(DataType const& dtype) noexcept(false)
     return names;
 }
 
-namespace details
-{
-
-// invalid, cannot convert to datatype
-template <typename ...Ts>
-struct datatype_impl
-{
-    // T must have an datatype associated!
-};
-
-// specialization for array types
-template <typename T, size_t N>
-struct datatype_impl<Array<T, N>>
-{
-    datatype_impl(CompoundNames<0> = {}) {}
-    operator DataType() const
-    {
-        return DataType::array(datatype_impl<T>(), N);
-    }
-};
-
-template <typename T, size_t N>
-struct datatype_impl<T[N]>
-{
-    datatype_impl(CompoundNames<0> = {}) {}
-    operator DataType() const
-    {
-        return datatype_impl<Array<T, N>>();
-    }
-};
-
-// specialization for varlen types
-template <typename T>
-struct datatype_impl<VarLen<T>>
-{
-    datatype_impl(CompoundNames<0> = {}) {}
-    operator DataType() const
-    {
-        return DataType::varLen(datatype_impl<T>());
-    }
-};
-
-// specialization for compound types
-template <typename... Ts>
-struct datatype_impl<Comp<Ts...>>
-{
-    using Compound  = conversion_t<Comp<Ts...>>;
-    using TypeNames = CompoundNames<sizeof...(Ts)>;
-
-    datatype_impl(TypeNames memberNames = {})
-        : m_typeNames(std::move(memberNames))
-    {
-        generateDefaultNames();
-    }
-
-    operator DataType() const
-    {
-        CompoundMembers members;
-        members.reserve(sizeof...(Ts));
-
-        mpl::static_for<sizeof...(Ts)>([&](auto const idx){
-            members.push_back({
-                m_typeNames.at(traits::comp_size<Compound>()-1-idx),
-                offset<idx>(),
-                datatype_impl<traits::rcomp_element_t<idx, Comp<Ts...>>>()
-            });
-        });
-        // members are flipped, not necessary for compound type creation,
-        // but makes debugging easier
-        std::reverse(std::begin(members), std::end(members));
-
-        return DataType::compound(sizeof(Compound), members);
-    }
-
-private:
-    /// compound member for calculating offset of an element
-    static Compound m_t;
-    /// names for compound members
-    TypeNames m_typeNames{};
-
-    /// offset of tuple element
-    template<size_t Idx>
-    size_t offset() const
-    {
-        return reinterpret_cast<size_t>(&get<Idx>(m_t)) -
-               reinterpret_cast<size_t>(&m_t);
-    }
-
-    void generateDefaultNames()
-    {
-        for (size_t i = 0; i < sizeof...(Ts); ++i)
-        {
-            if (m_typeNames[i].empty())
-            {
-                // String::number ambiguous for size_t
-                m_typeNames[i] = String{"type_"} + std::to_string(i);
-            }
-        }
-    }
-};
-
-// static compound_type init
-template <typename... Ts>
-conversion_t<Comp<Ts...>> datatype_impl<Comp<Ts...>>::m_t{};
-
-} // namespace details
-
-// compound types
-template<typename T1, typename T2, typename... Tother>
-inline DataType
-dataType(CompoundNames<sizeof...(Tother) + 2> memberNames = {}) noexcept(false)
-{
-    using T = Comp<T1, T2, Tother...>;
-    return details::datatype_impl<T>{std::move(memberNames)};
-}
-
-// compound names
-template<typename T>
-inline DataType
-dataType(CompoundNames<traits::comp_size<T>::value> memberNames) noexcept(false)
-{
-    return details::datatype_impl<T>{std::move(memberNames)};
-}
-
-template<typename T>
-inline DataType
-dataType() noexcept(false)
-{
-    return details::datatype_impl<T>{};
-}
-
 } // namespace GenH5
 
 // operators
@@ -351,48 +223,5 @@ swap(GenH5::DataType& a, GenH5::DataType& b) noexcept
 {
     a.swap(b);
 }
-
-#define GENH5_DECLARE_DATATYPE(NATIVE_TYPE, H5_TYPE) \
-    template <> \
-    struct GenH5::details::datatype_impl<NATIVE_TYPE> { \
-        datatype_impl(CompoundNames<0> = {}) {} \
-        operator GenH5::DataType() const { return H5_TYPE; } \
-    };
-
-#define GENH5_DECLARE_DATATYPE_IMPL(NATIVE_TYPE) \
-    template <> \
-    struct GenH5::details::datatype_impl<NATIVE_TYPE> { \
-        datatype_impl(CompoundNames<0> = {}) {} \
-        operator GenH5::DataType() const; \
-    }; \
-    inline \
-    GenH5::details::datatype_impl<NATIVE_TYPE>::operator GenH5::DataType() const
-
-// default datatypes
-GENH5_DECLARE_DATATYPE(bool, DataType::Bool);
-
-GENH5_DECLARE_DATATYPE(char, DataType::Char);
-
-GENH5_DECLARE_DATATYPE(char*, DataType::VarString);
-GENH5_DECLARE_DATATYPE(char const*, DataType::VarString);
-
-GENH5_DECLARE_DATATYPE(int, DataType::Int);
-GENH5_DECLARE_DATATYPE(long int, DataType::Long);
-GENH5_DECLARE_DATATYPE(long long int, DataType::LLong);
-GENH5_DECLARE_DATATYPE(unsigned int, DataType::UInt);
-GENH5_DECLARE_DATATYPE(unsigned long int, DataType::ULong);
-GENH5_DECLARE_DATATYPE(unsigned long long int, DataType::ULLong);
-
-GENH5_DECLARE_DATATYPE(float, DataType::Float);
-GENH5_DECLARE_DATATYPE(double, DataType::Double);
-
-GENH5_DECLARE_DATATYPE(GenH5::Version, DataType::Version);
-
-// fixed string
-template <size_t N>
-struct GenH5::details::datatype_impl<GenH5::Array<char, N>> {
-    datatype_impl(CompoundNames<0> = {}) {}
-    operator GenH5::DataType() const { return GenH5::DataType::string(N); }
-};
 
 #endif // GENH5_DATATYPE_H
